commands: Move validate functionality into package
